Full Implementation of the NPC Stumbling Mechanic
NPC's now can React to the Player Collision by either just stumbling or outright falling relative to the point of player to NPC contact, while also showering that knife-ear with a myriad of insults,  using the new and upcoming NPC Stumble mechanic!

Insults are pulled from the NPC Profanity Dex, which can be turned into a variety of Insult Profiles Tailormade for your NPC Needs.