Small Fixes
Fixed the Code added by the other 2 and integrated it better and added the GotHit to the player in the main scene.
Also made the Target allocation happen in the beginning, as the idea I originally had for it never really saw the light of day.
Intervening commit, because I want to save the adding GotHit to the scene before doing the tests with the enemies, as those can't be comitted.